feature : User, Schedule 권한 매핑

+++ src/main/java/com/example/schedule/schedule/service/ScheduleService.java
@@ -5,9 +5,12 @@ import com.example.schedule.schedule.dto.request.EditScheduleTitleAndContentsReq
 import com.example.schedule.schedule.dto.response.ScheduleResponse;
 import com.example.schedule.schedule.entity.Schedule;
 import com.example.schedule.schedule.repository.ScheduleRepository;
+import com.example.schedule.user.entity.User;
+import com.example.schedule.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.ObjectUtils;
 
 import java.util.List;
 
@@ -17,10 +20,13 @@ import java.util.List;
 public class ScheduleService {
 
     private final ScheduleRepository scheduleRepository;
+    private final UserRepository userRepository;
 
     @Transactional
-    public ScheduleResponse create(final CreateScheduleRequest request) {
-        final Schedule schedule = scheduleRepository.save(request.toEntity());
+    public ScheduleResponse create(final CreateScheduleRequest request, final Long loginId) {
+        User user = userRepository.findById(loginId)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않은 회원입니다."));
+        final Schedule schedule = scheduleRepository.save(request.toEntity(user));
         return ScheduleResponse.of(schedule);
     }
 
@@ -37,18 +43,34 @@ public class ScheduleService {
     }
 
     @Transactional
-    public ScheduleResponse editScheduleTitleAndContents(final Long id, final EditScheduleTitleAndContentsRequest request) {
+    public ScheduleResponse editScheduleTitleAndContents(final Long id, final EditScheduleTitleAndContentsRequest request, final Long loginId) {
+        User loginUser = userRepository.findById(loginId)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않은 회원입니다."));
+
         Schedule schedule = scheduleRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않은 일정입니다."));
+
+        if (!loginUser.isOwnerOf(schedule.getUser())) {
+            throw new IllegalArgumentException("권한이 없습니다.");
+        }
 
         schedule.updateTitleAndContents(request.getTitle(), request.getContents());
         return ScheduleResponse.of(schedule);
     }
 
     @Transactional
-    public void deleteById(Long id) {
+    public void deleteById(final Long id, final Long loginId) {
+        User loginUser = userRepository.findById(loginId)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않은 회원입니다."));
+
         Schedule schedule = scheduleRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않은 일정입니다."));
+
+        if (!loginUser.isOwnerOf(schedule.getUser())) {
+            throw new IllegalArgumentException("권한이 없습니다.");
+        }
+
         scheduleRepository.delete(schedule);
     }
+
 }
